cyclic xlink:href detection
