+++ engine/runtime/action-processor/src/feature/reader/citygml3/mod.rs
@@ -1,3 +1,5 @@
 pub(crate) mod geometry;
 pub(crate) mod parser;
 pub(crate) mod processor;
+pub(crate) mod utils;
+pub(crate) mod xlink;
